refactor: STORIF-335 - AccessKeysDrawer refactored.
