+++ packages/manager/src/features/ObjectStorage/AccessKeyLanding/AccessKeyLanding.tsx
@@ -1,26 +1,16 @@
-import {
-  createObjectStorageKeys,
-  revokeObjectStorageKey,
-  updateObjectStorageKey,
-} from '@linode/api-v4/lib/object-storage';
-import { useAccountSettings } from '@linode/queries';
 import { useErrors, useOpenClose } from '@linode/utilities';
 import { useNavigate } from '@tanstack/react-router';
 import * as React from 'react';
 
 import { DocumentTitleSegment } from 'src/components/DocumentTitle';
 import { PaginationFooter } from 'src/components/PaginationFooter/PaginationFooter';
-import { SecretTokenDialog } from 'src/features/Profile/SecretTokenDialog/SecretTokenDialog';
 import { usePaginationV2 } from 'src/hooks/usePaginationV2';
-import { useObjectStorageAccessKeys } from 'src/queries/object-storage/queries';
 import {
-  sendCreateAccessKeyEvent,
-  sendEditAccessKeyEvent,
-  sendRevokeAccessKeyEvent,
-} from 'src/utilities/analytics/customEventAnalytics';
-import { getAPIErrorOrDefault, getErrorMap } from 'src/utilities/errorUtils';
+  useDeleteAccessKeyMutation,
+  useObjectStorageAccessKeys,
+} from 'src/queries/object-storage/queries';
+import { getAPIErrorOrDefault } from 'src/utilities/errorUtils';
 
-import { useIsObjMultiClusterEnabled } from '../hooks/useIsObjectStorageGen2Enabled';
 import { AccessKeyDrawer } from './AccessKeyDrawer';
 import { AccessKeyTable } from './AccessKeyTable/AccessKeyTable';
 import { RevokeAccessKeyDialog } from './RevokeAccessKeyDialog';
@@ -30,9 +20,8 @@ import type { MODE, OpenAccessDrawer } from './types';
 import type {
   CreateObjectStorageKeyPayload,
   ObjectStorageKey,
-  UpdateObjectStorageKeyPayload,
 } from '@linode/api-v4/lib/object-storage';
-import type { FormikBag, FormikHelpers } from 'formik';
+import type { FormikBag } from 'formik';
 
 interface Props {
   accessDrawerOpen: boolean;
@@ -60,17 +49,11 @@ export const AccessKeyLanding = (props: Props) => {
     preferenceKey: 'object-storage-keys-table',
   });
 
-  const { data, error, isLoading, refetch } = useObjectStorageAccessKeys({
+  const { data, error, isLoading } = useObjectStorageAccessKeys({
     page: pagination.page,
     page_size: pagination.pageSize,
   });
-
-  const { data: accountSettings, refetch: requestAccountSettings } =
-    useAccountSettings();
-
-  // Key to display in Confirmation Modal upon creation
-  const [keyToDisplay, setKeyToDisplay] =
-    React.useState<null | ObjectStorageKey>(null);
+  const { mutateAsync: deleteAccessKey } = useDeleteAccessKeyMutation();
 
   // Key to rename (by clicking on a key's kebab menu )
   const [keyToEdit, setKeyToEdit] = React.useState<null | ObjectStorageKey>(
@@ -84,10 +67,7 @@ export const AccessKeyLanding = (props: Props) => {
   const [isRevoking, setIsRevoking] = React.useState<boolean>(false);
   const [revokeErrors, setRevokeErrors] = useErrors();
 
-  const displayKeysDialog = useOpenClose();
   const revokeKeysDialog = useOpenClose();
-
-  const { isObjMultiClusterEnabled } = useIsObjMultiClusterEnabled();
 
   // Redirect to base access keys route if current page has no data
   // TODO: Remove this implementation and replace `usePagination` with `usePaginate` hook. See [M3-10442]
@@ -108,123 +88,6 @@ export const AccessKeyLanding = (props: Props) => {
     }
   }, [data, isLoading, pagination.page, navigate]);
 
-  const handleCreateKey = (
-    values: CreateObjectStorageKeyPayload,
-    {
-      setErrors,
-      setStatus,
-      setSubmitting,
-    }: FormikHelpers<CreateObjectStorageKeyPayload>
-  ) => {
-    // Clear out status (used for general errors)
-    setStatus(null);
-    setSubmitting(true);
-
-    createObjectStorageKeys(values)
-      .then((data) => {
-        setSubmitting(false);
-
-        setKeyToDisplay(data);
-
-        // "Refresh" keys to include the newly created key
-        refetch();
-
-        props.closeAccessDrawer();
-        displayKeysDialog.open();
-
-        // If our Redux Store says that the user doesn't have OBJ enabled,
-        // it probably means they have just enabled it with the creation
-        // of this key. In that case, update the Redux Store so that
-        // subsequently created keys don't need to go through the
-        // confirmation flow.
-        if (accountSettings?.object_storage === 'disabled') {
-          requestAccountSettings();
-        }
-
-        // @analytics
-        sendCreateAccessKeyEvent();
-      })
-      .catch((errorResponse) => {
-        // We also need to refresh account settings on failure, since, depending
-        // on the error, Object Storage service might have actually been enabled.
-        if (accountSettings?.object_storage === 'disabled') {
-          requestAccountSettings();
-        }
-
-        setSubmitting(false);
-
-        const errors = getAPIErrorOrDefault(
-          errorResponse,
-          'There was an issue creating your Access Key.'
-        );
-        const mappedErrors = getErrorMap(['label'], errors);
-
-        // `status` holds general errors
-        if (mappedErrors.none) {
-          setStatus(mappedErrors.none);
-        }
-
-        setErrors(mappedErrors);
-      });
-  };
-
-  const handleEditKey = (
-    values: UpdateObjectStorageKeyPayload,
-    {
-      setErrors,
-      setStatus,
-      setSubmitting,
-    }: FormikHelpers<UpdateObjectStorageKeyPayload>
-  ) => {
-    // This shouldn't happen, but just in case.
-    if (!keyToEdit) {
-      return;
-    }
-
-    // Clear out status (used for general errors)
-    setStatus(null);
-
-    // If the new label is the same as the old one, no need to make an API
-    // request. Just close the drawer and return early.
-    if (values.label === keyToEdit.label) {
-      return closeAccessDrawer();
-    }
-
-    setSubmitting(true);
-
-    updateObjectStorageKey(
-      keyToEdit.id,
-      isObjMultiClusterEnabled ? values : { label: values.label }
-    )
-      .then((_) => {
-        setSubmitting(false);
-
-        // "Refresh" keys to display the newly updated key
-        refetch();
-
-        closeAccessDrawer();
-
-        // @analytics
-        sendEditAccessKeyEvent();
-      })
-      .catch((errorResponse) => {
-        setSubmitting(false);
-
-        const errors = getAPIErrorOrDefault(
-          errorResponse,
-          'There was an issue updating your Access Key.'
-        );
-        const mappedErrors = getErrorMap(['label'], errors);
-
-        // `status` holds general errors
-        if (mappedErrors.none) {
-          setStatus(mappedErrors.none);
-        }
-
-        setErrors(mappedErrors);
-      });
-  };
-
   const handleRevokeKeys = () => {
     // This shouldn't happen, but just in case.
     if (!keyToRevoke) {
@@ -234,17 +97,11 @@ export const AccessKeyLanding = (props: Props) => {
     setIsRevoking(true);
     setRevokeErrors([]);
 
-    revokeObjectStorageKey(keyToRevoke.id)
+    deleteAccessKey(keyToRevoke.id)
       .then((_) => {
         setIsRevoking(false);
 
-        // "Refresh" keys to remove the newly revoked key
-        refetch();
-
         revokeKeysDialog.close();
-
-        // @analytics
-        sendRevokeAccessKeyEvent();
       })
       .catch((errorResponse) => {
         setIsRevoking(false);
@@ -305,7 +162,6 @@ export const AccessKeyLanding = (props: Props) => {
         mode={mode}
         objectStorageKey={keyToEdit ? keyToEdit : undefined}
         onClose={closeAccessDrawer}
-        onSubmit={mode === 'creating' ? handleCreateKey : handleEditKey}
         open={accessDrawerOpen}
       />
 
@@ -314,12 +170,7 @@ export const AccessKeyLanding = (props: Props) => {
         onClose={closeAccessDrawer}
         open={mode === 'viewing' && accessDrawerOpen}
       />
-      <SecretTokenDialog
-        objectStorageKey={keyToDisplay}
-        onClose={displayKeysDialog.close}
-        open={displayKeysDialog.isOpen}
-        title="Access Keys"
-      />
+
       <RevokeAccessKeyDialog
         errors={revokeErrors}
         handleClose={closeRevokeDialog}
